CG-92: added verification for long and short packet formats.

+++ core/logging/messages.hpp
@@ -85,8 +85,8 @@ struct __attribute__((packed)) GPSLogEntry {
 			int32_t    velN;
 			int32_t    velE;
 			int32_t    velD;
-			int32_t    gSpeed;
-			double     headMot;
+			double     gSpeed;    // mm/s
+			double     headMot;   // Degrees
 			uint32_t   sAcc;
 			uint32_t   headAcc;
 			uint16_t   pDOP;
